refactor selectObjectContent to ts
