+++ src/internal/client.ts
@@ -14,6 +14,7 @@ import xml2js from 'xml2js'
 
 import { CredentialProvider } from '../CredentialProvider.ts'
 import * as errors from '../errors.ts'
+import type { SelectResults } from '../helpers.ts'
 import { DEFAULT_REGION, LEGAL_HOLD_STATUS, RETENTION_MODES, RETENTION_VALIDITY_UNITS } from '../helpers.ts'
 import { signV4 } from '../signing.ts'
 import { fsp, streamPromise } from './async.ts'
@@ -74,6 +75,7 @@ import type {
   ResponseHeader,
   ResultCallback,
   Retention,
+  SelectOptions,
   StatObjectOpts,
   Tag,
   TaggingOpts,
@@ -83,8 +85,13 @@ import type {
   VersionIdentificator,
 } from './type.ts'
 import type { ListMultipartResult, UploadedPart } from './xml-parser.ts'
+import {
+  parseCompleteMultipart,
+  parseInitiateMultipart,
+  parseObjectLegalHoldConfig,
+  parseSelectObjectContentResponse,
+} from './xml-parser.ts'
 import * as xmlParsers from './xml-parser.ts'
-import { parseCompleteMultipart, parseInitiateMultipart, parseObjectLegalHoldConfig } from './xml-parser.ts'
 
 const xml = new xml2js.Builder({ renderOpts: { pretty: false }, headless: true })
 
@@ -2210,5 +2217,77 @@ export class TypedClient {
     }
 
     await this.removeTagging({ bucketName, objectName, removeOpts })
+  }
+
+  async selectObjectContent(
+    bucketName: string,
+    objectName: string,
+    selectOpts: SelectOptions,
+  ): Promise<SelectResults | undefined> {
+    if (!isValidBucketName(bucketName)) {
+      throw new errors.InvalidBucketNameError(`Invalid bucket name: ${bucketName}`)
+    }
+    if (!isValidObjectName(objectName)) {
+      throw new errors.InvalidObjectNameError(`Invalid object name: ${objectName}`)
+    }
+    if (!_.isEmpty(selectOpts)) {
+      if (!isString(selectOpts.expression)) {
+        throw new TypeError('sqlExpression should be of type "string"')
+      }
+      if (!_.isEmpty(selectOpts.inputSerialization)) {
+        if (!isObject(selectOpts.inputSerialization)) {
+          throw new TypeError('inputSerialization should be of type "object"')
+        }
+      } else {
+        throw new TypeError('inputSerialization is required')
+      }
+      if (!_.isEmpty(selectOpts.outputSerialization)) {
+        if (!isObject(selectOpts.outputSerialization)) {
+          throw new TypeError('outputSerialization should be of type "object"')
+        }
+      } else {
+        throw new TypeError('outputSerialization is required')
+      }
+    } else {
+      throw new TypeError('valid select configuration is required')
+    }
+
+    const method = 'POST'
+    const query = `select&select-type=2`
+
+    const config: Record<string, unknown>[] = [
+      {
+        Expression: selectOpts.expression,
+      },
+      {
+        ExpressionType: selectOpts.expressionType || 'SQL',
+      },
+      {
+        InputSerialization: [selectOpts.inputSerialization],
+      },
+      {
+        OutputSerialization: [selectOpts.outputSerialization],
+      },
+    ]
+
+    // Optional
+    if (selectOpts.requestProgress) {
+      config.push({ RequestProgress: selectOpts?.requestProgress })
+    }
+    // Optional
+    if (selectOpts.scanRange) {
+      config.push({ ScanRange: selectOpts.scanRange })
+    }
+
+    const builder = new xml2js.Builder({
+      rootName: 'SelectObjectContentRequest',
+      renderOpts: { pretty: false },
+      headless: true,
+    })
+    const payload = builder.buildObject(config)
+
+    const res = await this.makeRequestAsync({ method, bucketName, objectName, query }, payload)
+    const body = await readAsBuffer(res)
+    return parseSelectObjectContentResponse(body)
   }
 }
